+++ src/index.ts
@@ -69,3 +69,4 @@ export {
 } from "./types"
 export { EventInfo } from "./events"
 export { safeWindow } from "./events/utils/window"
+export { AnimatePresence } from "./components/AnimatePresence"